// packages/users/src/user.interface.ts
import { type CreateUserDto } from "@platform/contracts";

export class UserEntity {
    id: number;
    name: string;
    email: string;
    password: string;
    createdAt: Date;
    updatedAt: Date;
}

export abstract class IUserService {
    abstract getUserById(id: number): Promise<UserEntity | null>;
    abstract getUserByEmail(email: string): Promise<UserEntity | null>;
    // getUserIdByToken(token: string): Promise<{ id: number }>;
    abstract createUser(user: CreateUserDto): Promise<UserEntity>;
    abstract updatePassword(userId: number, newPassword: string): Promise<{ id: number }>;
}